Bugfix/specific mindex (#35)
* fixed subblock matching to match on M. 
* added pixel_type function
* added an explanatory docstring note about the M Dimension and how it differs from m_index
* added pixel_type and M= unit test

+++ _aicspylibczi/SubblockMetaVec.h
@@ -45,21 +45,21 @@ namespace pylibczi {
           });
       }
 
-      std::vector<std::pair<char, int> >
+      std::vector<std::pair<char, size_t> >
       getShape(){
-          std::vector< std::map<char, int> > validIndexes;
+          std::vector< std::map<char, size_t> > validIndexes;
           for (const auto& image : *this) {
               validIndexes.push_back(image.getValidIndexes(m_isMosaic)); // only add M if it's a mosaic file
           }
 
           // TODO This code assumes the data is a matrix, meaning for example scene's have the same number of Z-slices
           // TODO is there another way to do this that could cope with variable data sizes within the matrix?
-          std::vector<std::pair<char, int> > charSizes;
-          std::map<char, std::set<int> > charSetSize;
-          std::map<char, std::set<int> >::iterator found;
+          std::vector<std::pair<char, size_t> > charSizes;
+          std::map<char, std::set<size_t> > charSetSize;
+          std::map<char, std::set<size_t> >::iterator found;
           for( const auto& validMap : validIndexes){
               for( auto keySet : validMap) {
-                  found = charSetSize.emplace(keySet.first, std::set<int>()).first;
+                  found = charSetSize.emplace(keySet.first, std::set<size_t>()).first;
                   found->second.insert(keySet.second);
               }
           }
@@ -67,7 +67,7 @@ namespace pylibczi {
               charSizes.emplace_back(keySet.first, keySet.second.size());
           }
           // sort them into descending DimensionIndex Order
-          std::sort(charSizes.begin(), charSizes.end(), [&](std::pair<char, int> a_, std::pair<char, int> b_){
+          std::sort(charSizes.begin(), charSizes.end(), [&](std::pair<char, size_t> a_, std::pair<char, size_t> b_){
               return libCZI::Utils::CharToDimension(a_.first) > libCZI::Utils::CharToDimension(b_.first);
           });
           return charSizes;
